+++ test/specs/openapi/test_schemas.py
@@ -11,7 +11,10 @@ from schemathesis.specs.openapi.parameters import OpenAPI20Parameter, OpenAPI30P
 @pytest.mark.operations("get_user", "update_user")
 def test_get_operation_via_remote_reference(openapi_version, schema_url):
     schema = schemathesis.from_uri(schema_url)
+    first = schema.get_operation_by_reference(f"{schema_url}#/paths/~1users~1{{user_id}}/patch")
     resolved = schema.get_operation_by_reference(f"{schema_url}#/paths/~1users~1{{user_id}}/patch")
+    # Check caching
+    assert resolved is first
     assert isinstance(resolved, APIOperation)
     assert resolved.path == "/users/{user_id}"
     assert resolved.method.upper() == "PATCH"
@@ -28,6 +31,33 @@ def test_get_operation_via_remote_reference(openapi_version, schema_url):
             "required": True,
             "schema": {"type": "integer"},
         }
+
+
+@pytest.mark.operations("get_user", "update_user")
+def test_operation_cache_sharing(mocker, schema_url):
+    # When the same operation is accessed via different methods
+    # The second access should use cache
+
+    def setup_mock(schema):
+        mocker.patch.object(
+            schema._operation_cache, "insert_operation_by_traversal_key", side_effect=ValueError("Not cached")
+        )
+
+    reference = f"{schema_url}#/paths/~1users~1{{user_id}}/patch"
+    operation_id = "updateUser"
+
+    schema = schemathesis.from_uri(schema_url)
+    first = schema.get_operation_by_reference(reference)
+    # After accessing by reference, there should not be an attempt to insert it again
+    setup_mock(schema)
+    second = schema.get_operation_by_id(operation_id)
+    assert first is second
+    # And the other way around
+    schema = schemathesis.from_uri(schema_url)
+    first = schema.get_operation_by_id(operation_id)
+    setup_mock(schema)
+    second = schema.get_operation_by_reference(reference)
+    assert first is second
 
 
 SINGLE_METHOD_PATHS = {